support panging on geting result api and revise code
