+++ backend/components/v2/admin/index.js
@@ -35,7 +35,15 @@ const users = require('./user');
 const uuidv4 = require('uuid/v4');
 const auth = require('../../helpers/auth');
 const members = require('./members');
-
+const {saveSurvey,
+    createEmptySurvey,
+    getAllSurveysByOwnerId,
+    getSurveyById,
+    renameSurvey,
+    deleteSurvey,
+    getAllResultsBySurveyId,
+    getResultById,
+    saveResult} = require('./survey');
 
 router.use(function (req, res, next) {
     logger.info('calling users api ' + req.path);
@@ -93,331 +101,6 @@ router.use(function (req, res, next) {
         }
     })
 });
-
-/** Survey Management Services */
-
-function saveSurvey(req, res) {
-
-    console.log(req.body)
-
-    var body = req.body
-
-    var searchFilters = {
-        userid: body.userid,
-        surveyid: body.surveyid,
-        version: body.version,
-    };
-
-    let resultProvmises = mongo.find(searchFilters, appConf.surveyCollections.survey, {});
-
-    var survey = {
-        userid: body.userid,
-        name: body.name,
-        surveyid: body.surveyid,
-        version: body.version,
-        pages: body.pages,
-        created_at: new Date(),
-        modified_at: new Date(),
-    }
-
-    var filters = {
-        userid: body.userid,
-        name: body.name,
-        surveyid: body.surveyid,
-        version: body.version
-    };
-
-    resultProvmises.then(results => {
-            return Promise.resolve(results);
-        })
-        .catch(err => {
-            mongo.insert(filters, survey, appConf.surveyCollections.survey).then(result => {
-                    http.success(res, result);
-                    return true;
-                })
-                .catch(err => {
-                    http.error(res, 500, 50002, "mongo error: " + err);
-                    return false;
-                });
-        })
-        .then(results => {
-            if (results[0] !== undefined && results[0].created_at !== undefined) {
-                console.log("created at : " + results[0].created_at);
-                survey.created_at = results[0].created_at;
-            }
-            mongo.insert(filters, survey, appConf.surveyCollections.survey).then(result => {
-                    http.success(res, result);
-                    return true;
-                })
-                .catch(err => {
-                    http.error(res, 500, 50002, "mongo error: " + err);
-                    return false;
-                });
-        })
-
-
-}
-
-function createEmptySurvey(req, res) {
-    console.log(req.body)
-
-    var body = req.body
-
-    if (body.userid === undefined) {
-        http.error(res, 400, 40000, 'require "userid"');
-        return;
-    }
-
-    if (body.name === undefined) {
-        http.error(res, 400, 40000, 'require "name"');
-        return;
-    }
-
-    if (body.surveyid === undefined) {
-        http.error(res, 400, 40000, 'require "surveyid"');
-        return;
-    }
-
-    if (body.version === undefined) {
-        http.error(res, 400, 40000, 'require "version"');
-        return;
-    }
-
-
-    var survey = {
-        userid: body.userid,
-        name: body.name,
-        surveyid: body.surveyid,
-        version: body.version,
-        created_at: new Date(),
-        modified_at: new Date(),
-    }
-
-    var filters = {
-        userid: body.userid,
-        name: body.name,
-        surveyid: body.surveyid,
-        version: body.version
-    };
-
-    mongo.insert(filters, survey, appConf.surveyCollections.survey).then(result => {
-            http.success(res, result);
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-}
-
-function getAllSurveysByOwnerId(req, res) {
-    if (req.params.ownerid == 'undefined') {
-        http.error(res, 400, 40000, "Not found owerid");
-        return false;
-    }
-
-    var filters = {
-        userid: req.params.ownerid
-    };
-
-    var sort = {
-        created_at: -1
-    };
-
-    mongo.find(filters, appConf.surveyCollections.survey, sort).then(results => {
-            http.success(res, results);
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-
-}
-
-function getSurveyById(req, res) {
-
-    var userid = 'undefined';
-    var version = '1';
-
-    if (req.query.uid === undefined) {
-        http.error(res, 400, 40000, "Not found userid");
-        return;
-    } else {
-        userid = req.query.uid + "";
-    }
-
-    if (req.query.v !== undefined) {
-        version = req.query.v + "";
-    }
-
-    if (req.params.surveyId === undefined) {
-        http.error(res, 400, 40000, "Not found owerid");
-        return;
-    }
-
-    var filters = {
-        surveyid: req.params.surveyId,
-        version: version,
-        userid: userid,
-    };
-
-    mongo.find(filters, appConf.surveyCollections.survey).then(results => {
-            http.success(res, results[0]);
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-};
-
-function renameSurvey(req, res) {
-
-    var filters = {
-        surveyid: req.body.surveyid,
-        userid: req.body.userid,
-        version: req.body.version
-    };
-
-
-    var data = {
-        $set: {
-            name: req.body.name,
-            modified_at: new Date(),
-        }
-    };
-
-    mongo.update(filters, data, appConf.surveyCollections.survey).then(results => {
-            http.success(res, results);
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-}
-
-function deleteSurvey(req, res) {
-    var filters = {
-        userid: req.body.userid,
-        surveyid: req.body.surveyid,
-        version: req.body.version
-    };
-
-    mongo.remove(filters, true, appConf.surveyCollections.survey).then(result => {
-            http.success(res, {
-                affected: result.result.n
-            });
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-}
-
-/** Answer Management Services */
-function getAllResultsBySurveyId(req, res) {
-
-    var userid = 'undefined';
-    var version = '1';
-
-    if (req.query.uid !== undefined) {
-        userid = req.query.uid;
-    }
-
-    if (req.query.v !== undefined) {
-        version = req.query.v;
-    }
-
-    var filters = {
-        surveyid: req.params.surveyId + "",
-        userid: userid,
-        version: version
-    };
-
-    var sort = {
-        created_at: -1
-    };
-
-    mongo.find(filters, appConf.surveyCollections.result, sort).then(results => {
-            http.success(res, results);
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-
-}
-
-function getResultById(req, res) {
-
-    var body = req.body;
-
-    var filters = {
-        resultid: req.params.resultId
-    }
-
-    mongo.find(filters, appConf.surveyCollections.result).then(results => {
-            http.success(res, results);
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-}
-
-function saveResult(req, res) {
-
-    var body = req.body;
-    var resultid = 'undefined';
-    var userid = 'undefined';
-    var version = '1';
-
-    if (body.resultid === undefined || body.resultid == null) {
-        resultid = uuidv4();
-    } else {
-        resultid = body.resultid;
-    }
-
-    if (body.userid !== undefined) {
-        userid = body.userid;
-    }
-
-    if (body.version !== undefined) {
-        version = body.version;
-    }
-
-    var filters = {
-        surveyid: body.surveyid,
-        resultid: resultid,
-        version: version,
-        userid: userid,
-
-    }
-
-    var data = {
-        surveyid: body.surveyid,
-        resultid: resultid,
-        version: version,
-        userid: userid,
-        result: body.result,
-        created_at: new Date(),
-        modified_at: new Date(),
-    }
-
-    mongo.insert(filters, data, appConf.surveyCollections.result).then(result => {
-            http.success(res, result);
-            return true;
-        })
-        .catch(err => {
-            http.error(res, 500, 50002, "mongo error: " + err);
-            return false;
-        });
-}
 
 // Survey Management Services
 // /v2/survey/:userID
